backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0,<11.0
python-decouple>=3.8,<4.0
gunicorn>=21.0,<23.0
whitenoise>=6.0,<7.0
